deleting obsolete code
