+++ command/test_command_file_changing_flows.py
@@ -668,197 +668,6 @@ class TestFileChangingFlows(IsolatedAsyncioTestCase):
             dump_file_list(self.tmpdir.name + "/repo-full/wat", "", data=True),
             dump_file_list(self.tmpdir.name + "/new-contents", "", data=True))
 
-    async def test_resetting_file_contents(self):
-        hoard_cmd, partial_cave_cmd, full_cave_cmd, backup_cave_cmd, incoming_cave_cmd = \
-            await init_complex_hoard(self.tmpdir.name)
-
-        pfw = pretty_file_writer(self.tmpdir.name)
-        os.mkdir(join(self.tmpdir.name, "changed-cave"))
-        pfw('changed-cave/test.me.1', "w3q45yhq3g")
-        pfw('changed-cave/test.me.4', "fwadeaewdsa")
-        pfw('changed-cave/wat/test.me.2', "gsadf3dq")
-        pfw('changed-cave/wat/test.me.3', "'psejmfw'")
-
-        changed_cave_cmd = TotalCommand(path=join(self.tmpdir.name, "changed-cave")).cave
-        changed_cave_cmd.init()
-        res = await changed_cave_cmd.refresh(show_details=False)
-        self.assertEqual('old: None\ncurrent: 44be92d978ad301ad4463c09d88ed9b0d828cb4a\nRefresh done!', res)
-
-        res = hoard_cmd.add_remote(
-            remote_path=join(self.tmpdir.name, "changed-cave"), name="repo-changed-cave-name",
-            mount_point="/", type=CaveType.PARTIAL, fetch_new=False)
-        self.assertEqual(
-            fr"Added repo-changed-cave-name[{changed_cave_cmd.current_uuid()}] at {self.tmpdir.name}\changed-cave!",
-            res)
-
-        res = await hoard_cmd.contents.pull(partial_cave_cmd.current_uuid())
-        self.assertEqual((
-            'Pulling repo-partial-name...\n'
-            'Before: Hoard [a80f91] <- repo [curr: a80f91, stg: f9bfc2, des: a80f91]\n'
-            'REPO_MARK_FILE_AVAILABLE /test.me.1\n'
-            'HOARD_FILE_ADDED /test.me.1\n'
-            'REPO_MARK_FILE_AVAILABLE /wat/test.me.2\n'
-            'HOARD_FILE_ADDED /wat/test.me.2\n'
-            'updated repo-partial-name from a80f91 to f9bfc2\n'
-            'updated repo-full-name from a80f91 to f9bfc2\n'
-            'updated repo-backup-name from a80f91 to f9bfc2\n'
-            'After: Hoard [f9bfc2], repo [curr: f9bfc2, stg: f9bfc2, des: f9bfc2]\n'
-            "Sync'ed repo-partial-name to hoard!\n"
-            'DONE'), res)
-
-        res = await hoard_cmd.contents.pull(full_cave_cmd.current_uuid())
-        self.assertEqual((
-            'Pulling repo-full-name...\n'
-            'Before: Hoard [f9bfc2] <- repo [curr: a80f91, stg: 1ad9e0, des: f9bfc2]\n'
-            'REPO_MARK_FILE_AVAILABLE /test.me.1\n'
-            'REPO_MARK_FILE_AVAILABLE /test.me.4\n'
-            'HOARD_FILE_ADDED /test.me.4\n'
-            'REPO_MARK_FILE_AVAILABLE /wat/test.me.2\n'
-            'REPO_MARK_FILE_AVAILABLE /wat/test.me.3\n'
-            'HOARD_FILE_ADDED /wat/test.me.3\n'
-            'updated repo-full-name from f9bfc2 to 1ad9e0\n'
-            'updated repo-backup-name from f9bfc2 to 1ad9e0\n'
-            'After: Hoard [1ad9e0], repo [curr: 1ad9e0, stg: 1ad9e0, des: 1ad9e0]\n'
-            "Sync'ed repo-full-name to hoard!\n"
-            'DONE'), res)
-
-        res = await hoard_cmd.contents.status(hide_time=True, hide_disk_sizes=True)
-        self.assertEqual(
-            'Root: 1ad9e0f92a8411689b1aee57f9ccf36c1f09a1ad\n'
-            "|Num Files                |total     |available |get       |\n"
-            "|repo-backup-name         |         4|          |         4|\n"
-            "|repo-full-name           |         4|         4|          |\n"
-            "|repo-partial-name        |         2|         2|          |\n"
-            "\n"
-            "|Size                     |total     |available |get       |\n"
-            "|repo-backup-name         |        35|          |        35|\n"
-            "|repo-full-name           |        35|        35|          |\n"
-            "|repo-partial-name        |        14|        14|          |\n", res)
-
-        res = await hoard_cmd.contents.pull(changed_cave_cmd.current_uuid(), assume_current=True)
-        self.assertEqual((
-            'Pulling repo-changed-cave-name...\n'
-            'Before: Hoard [1ad9e0] <- repo [curr: a80f91, stg: 44be92, des: a80f91]\n'
-            'REPO_MARK_FILE_AVAILABLE /test.me.1\n'
-            'HOARD_FILE_CHANGED /test.me.1\n'
-            'REPO_MARK_FILE_AVAILABLE /test.me.4\n'
-            'REPO_MARK_FILE_AVAILABLE /wat/test.me.2\n'
-            'REPO_MARK_FILE_AVAILABLE /wat/test.me.3\n'
-            'HOARD_FILE_CHANGED /wat/test.me.3\n'
-            'updated repo-partial-name from f9bfc2 to 0bf8fb\n'
-            'updated repo-full-name from 1ad9e0 to 44be92\n'
-            'updated repo-backup-name from 1ad9e0 to 44be92\n'
-            'updated repo-changed-cave-name from a80f91 to 44be92\n'
-            'After: Hoard [44be92], repo [curr: 44be92, stg: 44be92, des: 44be92]\n'
-            "Sync'ed repo-changed-cave-name to hoard!\n"
-            'DONE'), res)
-
-        res = await hoard_cmd.contents.status(hide_time=True, hide_disk_sizes=True)
-        self.assertEqual(
-            'Root: 44be92d978ad301ad4463c09d88ed9b0d828cb4a\n'
-            "|Num Files                |total     |available |get       |\n"
-            "|repo-backup-name         |         4|          |         4|\n"
-            "|repo-changed-cave-name   |         4|         4|          |\n"
-            "|repo-full-name           |         4|         2|         2|\n"
-            "|repo-partial-name        |         2|         1|         1|\n"
-            "\n"
-            "|Size                     |total     |available |get       |\n"
-            "|repo-backup-name         |        38|          |        38|\n"
-            "|repo-changed-cave-name   |        38|        38|          |\n"
-            "|repo-full-name           |        38|        19|        19|\n"
-            "|repo-partial-name        |        18|         8|        10|\n", res)
-
-        res = await hoard_cmd.files.pending()
-        self.assertEqual(
-            "repo-partial-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            " repo-changed-cave-name has 1 files\n"
-            "repo-full-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            "TO_GET (from 1) /wat/test.me.3\n"
-            " repo-changed-cave-name has 2 files\n"
-            "repo-backup-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            "TO_GET (from 2) /test.me.4\n"
-            "TO_GET (from 3) /wat/test.me.2\n"
-            "TO_GET (from 1) /wat/test.me.3\n"
-            " repo-changed-cave-name has 4 files\n"
-            " repo-full-name has 2 files\n"
-            " repo-partial-name has 1 files\n"
-            "repo-incoming-name:\n"
-            "repo-changed-cave-name:\n"
-            "DONE", res)
-
-        # resetting pending ops
-        res = await hoard_cmd.contents.reset("repo-full-name")
-        self.assertEqual(
-            "repo-full-name:\n"
-            "WONT_GET /test.me.1\n"
-            "WONT_GET /wat/test.me.3\n"
-            "DONE", res)
-
-        res = await hoard_cmd.files.pending()
-        self.assertEqual(
-            "repo-partial-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            " repo-changed-cave-name has 1 files\n"
-            "repo-full-name:\n"
-            "repo-backup-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            "TO_GET (from 2) /test.me.4\n"
-            "TO_GET (from 3) /wat/test.me.2\n"
-            "TO_GET (from 1) /wat/test.me.3\n"
-            " repo-changed-cave-name has 4 files\n"
-            " repo-full-name has 2 files\n"
-            " repo-partial-name has 1 files\n"
-            "repo-incoming-name:\n"
-            "repo-changed-cave-name:\n"
-            "DONE", res)
-
-        # resetting existing contents to what repo-full-name should contain!
-        res = await hoard_cmd.contents.reset_with_existing("repo-full-name")
-        self.assertEqual(
-            "repo-full-name:\n"
-            "RESET /test.me.1\n"
-            "RESET /wat/test.me.3\n"
-            "DONE", res)
-
-        res = await hoard_cmd.files.pending()
-        self.assertEqual(
-            "repo-partial-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            " repo-full-name has 1 files\n"
-            "repo-full-name:\n"
-            "repo-backup-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            "TO_GET (from 2) /test.me.4\n"
-            "TO_GET (from 3) /wat/test.me.2\n"
-            "TO_GET (from 1) /wat/test.me.3\n"
-            " repo-changed-cave-name has 2 files\n"
-            " repo-full-name has 4 files\n"
-            " repo-partial-name has 1 files\n"
-            "repo-incoming-name:\n"
-            "repo-changed-cave-name:\n"
-            "TO_GET (from 1) /test.me.1\n"
-            "TO_GET (from 1) /wat/test.me.3\n"
-            " repo-full-name has 2 files\n"
-            "DONE", res)
-
-        res = await hoard_cmd.contents.status(hide_time=True, hide_disk_sizes=True)
-        self.assertEqual(
-            'Root: 44be92d978ad301ad4463c09d88ed9b0d828cb4a\n'
-            "|Num Files                |total     |available |get       |\n"
-            "|repo-backup-name         |         4|          |         4|\n"
-            "|repo-changed-cave-name   |         4|         2|         2|\n"
-            "|repo-full-name           |         4|         4|          |\n"
-            "|repo-partial-name        |         2|         1|         1|\n"
-            "\n"
-            "|Size                     |total     |available |get       |\n"
-            "|repo-backup-name         |        35|          |        35|\n"
-            "|repo-changed-cave-name   |        35|        19|        16|\n"
-            "|repo-full-name           |        35|        35|          |\n"
-            "|repo-partial-name        |        14|         8|         6|\n", res)
-
     async def test_moving_of_files_in_hoard(self):
         hoard_cmd, partial_cave_cmd, full_cave_cmd, backup_cave_cmd, incoming_cave_cmd = \
             await init_complex_hoard(self.tmpdir.name)
